requirements.txt
tensorflow
pillow
numpy
pandas
matplotlib
scikit-learn
streamlit
opencv-python